Change maxdiff from int64_t to double for fractional diff support
Mirrors the mindiff/startdiff type-change pattern. maxdiff now accepts
floating-point values in ckpool.conf, enabling fine-grained difficulty
caps for low-hashrate or testing environments.

Values greater than 1 are floored to the nearest integer at load time.
Zero (or any value less than 1) preserves the existing "no maximum"
behaviour; the stratifier guard was updated from `if (ckp->maxdiff)` to
`if (ckp->maxdiff >= 1.0)` to make this explicit and to correctly
exclude sub-1 fractional values from acting as a cap.

Uses the existing json_get_num_as_double() helper so both integer and
real JSON values (e.g. `"maxdiff" : 0` and `"maxdiff" : 500000.0`)
continue to parse correctly.

Changes:
- src/ckpool.h: maxdiff int64_t -> double
- src/ckpool.c: parse maxdiff with json_get_num_as_double(); apply
  floor() for values > 1 in main()
- src/stratifier.c: update guard to >= 1.0; add (int64_t) cast in
  MIN macro
- README: update maxdiff description to document float support and
  sub-1 = no maximum semantics

+++ src/ckpool.h
@@ -236,7 +236,7 @@ struct ckpool_instance {
 	int64_t mindiff; // Default 1
 	int64_t startdiff; // Default 42
 	int64_t highdiff; // Default 1000000
-	int64_t maxdiff; // No default
+	double maxdiff; // No default (0 = unlimited)
 
 	/* Coinbase data */
 	char *btcaddress; // Address to mine to
